analytics/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
motor
pandas
